requirements.txt
asgiref==3.2.3
beautifulsoup4==4.8.2
bootstrap4==0.1.0
Django==3.0.3
django-bootstrap4==1.1.1
django-ckeditor==5.9.0
django-debug-toolbar==2.2
django-js-asset==1.2.2
mysqlclient==1.4.6
Pillow==7.0.0
pytz==2019.3
soupsieve==2.0
sqlparse==0.3.0